% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_results}
\alias{available_results}
\title{Extract final configurations into a data frame}
\usage{
available_results(obj)
}
\arguments{
\item{obj}{An object of class \code{scregclust}}
}
\value{
A \code{\link{data.frame}} containing penalization parameters and
final configurations for those penalizations.
}
\description{
Extract final configurations into a data frame
}
\concept{helpers}
