% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_overlap}
\alias{cluster_overlap}
\title{Create a table of module overlap for two clusterings}
\usage{
cluster_overlap(k1, k2)
}
\arguments{
\item{k1}{First clustering}

\item{k2}{Second clustering}
}
\value{
A matrix showing the module overlap with the labels of \code{k1} in
the columns and the labels of \code{k2} in the rows.
}
\description{
Compares two clusterings and creates a table of overlap between them.
Module labels do not have to match.
}
\concept{helpers}
