% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef_ols}
\alias{coef_ols}
\title{Compute OLS coefficients}
\usage{
coef_ols(y, x)
}
\arguments{
\item{y}{Target vector (n x 1)/matrix (n x m)}

\item{x}{Design matrix (n x p)}
}
\value{
Vector of OLS coefficients
}
\description{
If the design matrix has full column-rank, then use the normal
least squares estimate. Otherwise, use the Moore-Penrose inverse
to compute the least squares estimate.
}
\keyword{internal}
