% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef_ridge}
\alias{coef_ridge}
\title{Compute ridge regression coefficients}
\usage{
coef_ridge(y, x, lambda)
}
\arguments{
\item{y}{Target vector (n x 1)/matrix (n x m)}

\item{x}{Design matrix (n x p)}

\item{lambda}{Positive parameter for ridge penalty}
}
\value{
Vector of ridge regression coefficients
}
\description{
Compute ridge regression coefficients
}
\keyword{internal}
