% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_adjusted_rand_index}
\alias{compute_adjusted_rand_index}
\title{Compute Hubert's and Arabie's Adjusted Rand index}
\usage{
compute_adjusted_rand_index(k1, k2)
}
\arguments{
\item{k1}{First clustering as vector of integers}

\item{k2}{Second clustering as vector of integers}
}
\value{
The Adjusted Rand index as a numeric value
}
\description{
Compute Hubert's and Arabie's Adjusted Rand index
}
\references{
Lawrence Hubert and Phipps Arabie (1985). "Comparing partitions".
Journal of Classification. 2 (1): 193–218. DOI:10.1007/BF01908075
}
\keyword{internal}
