% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_rand_index}
\alias{compute_rand_index}
\title{Compute the Rand index}
\usage{
compute_rand_index(k1, k2)
}
\arguments{
\item{k1}{First clustering as vector of integers}

\item{k2}{Second clustering as vector of integers}
}
\value{
The Rand index as a numeric value
}
\description{
Compute the Rand index
}
\references{
W. M. Rand (1971). "Objective criteria for the evaluation of clustering
methods". Journal of the American Statistical Association 66 (336): 846–850.
DOI:10.2307/2284239
}
\keyword{internal}
