% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fast_cor}
\alias{fast_cor}
\title{Fast computation of correlation}
\usage{
fast_cor(x, y)
}
\arguments{
\item{x}{first input matrix}

\item{y}{second input matrix}
}
\value{
Correlations matrix between the columns of \code{x} and \code{y}
}
\description{
This uses a more memory-intensive but much faster algorithm than
the built-in \code{cor} function.
}
\details{
Computes the correlation between the columns of \code{x} and \code{y}.
}
\concept{helpers}
