% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_cluster_sizes}
\alias{find_cluster_sizes}
\title{Determine cluster sizes}
\usage{
find_cluster_sizes(k, n_cl)
}
\arguments{
\item{k}{Vector of cluster indices}

\item{n_cl}{Total number of clusters}
}
\value{
A named vector containining the name of the cluster (its index or
\code{"Noise"}) and the number of elements in that cluster
}
\description{
Determine cluster sizes
}
\concept{helpers}
