% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_module_sizes}
\alias{find_module_sizes}
\title{Determine module sizes}
\usage{
find_module_sizes(module, n_modules)
}
\arguments{
\item{module}{Vector of module indices}

\item{n_modules}{Total number of modules}
}
\value{
A named vector containining the name of the module (its index or
\code{"Noise"}) and the number of elements in that module
}
\description{
Determine module sizes
}
\concept{helpers}
