% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_avg_num_regulators}
\alias{get_avg_num_regulators}
\title{Get the average number of active regulators per module}
\usage{
get_avg_num_regulators(fit)
}
\arguments{
\item{fit}{An object of class \code{scRegClust}}
}
\value{
A \code{\link{data.frame}} containing the average number of active regulators
per module for each penalization parameter.
}
\description{
Get the average number of active regulators per module
}
\concept{utilities}
