% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_final_configs}
\alias{get_num_final_configs}
\title{Return the number of final configurations}
\usage{
get_num_final_configs(fit)
}
\arguments{
\item{fit}{An object of class \code{scRegClust}}
}
\value{
An integer vector containing the number of final configurations
for each penalization parameter.
}
\description{
Returns the number of final configurations per penalization parameter in an
scRegClust object.
}
\concept{utilities}
