% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scregclust.R
\name{get_regulator_list}
\alias{get_regulator_list}
\title{Return list of regulator genes}
\usage{
get_regulator_list(mode = c("TF", "kinase"))
}
\arguments{
\item{mode}{Determines which genes are considered to be regulators.
Currently supports TF=transcription factors and kinases.}
}
\value{
a list of gene symbols
}
\description{
Return list of regulator genes
}
\seealso{
\code{\link[=scregclust_format]{scregclust_format()}}
}
\concept{utilities}
