% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kmeanspp_init}
\alias{kmeanspp_init}
\title{Determine initial centers for the kmeans++ algorithm}
\usage{
kmeanspp_init(n_cluster, x = NULL, dm = NULL)
}
\arguments{
\item{x}{data matrix to be clustered}

\item{dm}{distance matrix (between rows of x; of class "dist")}
}
\value{
Row indices of initial cluster centres of x
}
\description{
Determine initial centers for the kmeans++ algorithm
}
\keyword{internal}
