% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_module_count_helper}
\alias{plot_module_count_helper}
\title{Plot average silhouette scores and average predictive \eqn{R^2}}
\usage{
plot_module_count_helper(list_of_fits, penalization)
}
\arguments{
\item{list_of_fits}{A list of \code{scregclust} objects each fit to the same
dataset across a variety of module counts (varying
\code{n_modules} while running \code{\link{scregclust}}).}

\item{penalization}{Either a single numeric value requesting the results
for the same penalty parameter across all fits in
\code{list_of_fits}, or one for each individual fit.}
}
\value{
A ggplot2 plot showing the average silhouette score and the
average predictive \eqn{R^2}
}
\description{
Plot average silhouette scores and average predictive \eqn{R^2}
}
\concept{plotting}
