% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_silhouettes}
\alias{plot_silhouettes}
\title{Plot individual silhouette scores}
\usage{
plot_silhouettes(list_of_fits, penalization, final_config = 1L)
}
\arguments{
\item{list_of_fits}{A list of \code{scregclust} objects each fit to the same
dataset across a variety of module counts (varying
\code{n_modules} when running \code{\link{scregclust}}).}

\item{penalization}{Either a single numeric value requesting the results
for the same penalty parameter across all fits in
\code{list_of_fits}, or one for each individual fit.}

\item{final_config}{The final configuration that should be visualized.
Either a single number to be used for all fits in
\code{list_of_fits}, or one for each individual fit.}
}
\value{
A ggplot2 plot showing the the silhouette scores for each
supplied fit.
}
\description{
Plot individual silhouette scores
}
\concept{plotting}
