% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progstr}
\alias{progstr}
\title{Quick'n'dirty progress bar}
\usage{
progstr(step, n_steps, name, finished = FALSE, progress_length = 20L)
}
\arguments{
\item{step}{current step being worked on}

\item{n_steps}{total number of steps}

\item{name}{name of the process}

\item{finished}{whether the process is finished}

\item{progress_length}{length of the progress bar in ascii signs}
}
\value{
A string formatted as a progress bar
}
\description{
Creates a progress bar and returns it as a string.
}
\keyword{internal}
