% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_modules}
\alias{remove_empty_modules}
\title{Remove empty modules}
\usage{
remove_empty_modules(module)
}
\arguments{
\item{module}{Vector of module indices}
}
\value{
The updated vector of module indices with empty modules removed.
}
\description{
Remove empty modules
}
\details{
Only iterates through modules with positive index, leaving the noise
module untouched.
}
\keyword{internal}
