% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reset_array}
\alias{reset_array}
\title{Reset input 3d-array by filling matrix along first dimension}
\usage{
reset_array(arr, input)
}
\arguments{
\item{arr}{The 3d-array of dimension \verb{n_cl x n_obs x n_genes}}

\item{input}{The matrix of size \verb{n_obs x n_genes}}
}
\description{
Reset input 3d-array by filling matrix along first dimension
}
\keyword{internal}
