% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scregclust.R
\name{scregclust_format}
\alias{scregclust_format}
\title{Package data before clustering}
\usage{
scregclust_format(expression_matrix, mode = c("TF", "kinase"))
}
\arguments{
\item{expression_matrix}{The p x n gene expression matrix with gene symbols
as rownames.}

\item{mode}{Determines which genes are considered to be regulators.}
}
\value{
A list with
\item{genesymbols}{The gene symbols extracted from the expression matrix}
\item{sample_assignment}{A vector filled with \code{1}'s of the same length as
there are columns in the gene expression matrix.}
\item{is_regulator}{Whether a gene is considered to be a regulator or not,
determined dependent on \code{mode}.}
}
\description{
Package data before clustering
}
\seealso{
\code{\link[=get_regulator_list]{get_regulator_list()}}
}
\concept{main}
