% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_cluster_indices}
\alias{update_cluster_indices}
\title{Remove empty clusters}
\usage{
update_cluster_indices(k)
}
\arguments{
\item{k}{Vector of cluster indices}
}
\value{
The updated vector of cluster indices with empty clusters removed.
}
\description{
Remove empty clusters
}
\details{
Only iterates through clusters with positive index, leaving the noise
cluster untouched.
}
\keyword{internal}
