% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scregclust-package.R
\docType{package}
\name{scregclust-package}
\alias{scregclust-package}
\title{scregclust: Reconstructing the regulatory programs of target genes in single cell data}
\description{
This package provides an implementation of our scRegClust algorithm (Larsson, Held, et al., 2023, \doi{10.1101/2023.03.10.532041}) which aims to reconstruct the regulatory programs of target genes in single cell data. Target genes are clustered and each cluster is associated with a linear model accounting for the regulatory program driving the genes in the cluster.
}
\details{
Computational methods for the scregclust algorithm
}
\author{
Ida Larsson, Felix Held, Sven Nelander
}
\keyword{internal}
