% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scregclust-package.R
\docType{package}
\name{scregclust-package}
\alias{scregclust-package}
\title{scregclust: Reconstructing the Regulatory Programs of Target Genes in scRNA-Seq Data}
\description{
Implementation of the scregclust algorithm described in Larsson, Held, et al. (2024) \doi{10.1038/s41467-024-53954-3} which reconstructs regulatory programs of target genes in scRNA-seq data. Target genes are clustered into modules and each module is associated with a linear model describing the regulatory program.
}
\details{
Computational methods for the scregclust algorithm
}
\seealso{
Useful links:
\itemize{
  \item \url{https://scmethods.github.io/scregclust/}
  \item \url{https://github.com/scmethods/scregclust/}
  \item Report bugs at \url{https://github.com/scmethods/scregclust/issues}
}

}
\author{
Ida Larsson, Felix Held, Sven Nelander
}
\keyword{internal}
